refactor(step5): view 패키지 이동
