+++ src/main/kotlin/study/racingcar/view/InputView.kt
@@ -1,4 +1,4 @@
-package study.racingcar
+package study.racingcar.view
 
 object InputView {
