+++ src/main/kotlin/study/racingcar/view/ResultView.kt
@@ -1,4 +1,6 @@
-package study.racingcar
+package study.racingcar.view
+
+import study.racingcar.domain.Car
 
 object ResultView {
